requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Django==5.0.3
django-radius==1.5.1
fonttools==4.50.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.8.4
mysql-connector==2.2.9
mysql-connector-python==8.3.0
netaddr==1.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
PyMySQL==1.1.0
pyparsing==3.1.2
pyrad==2.4
pytesseract==0.3.10
python-dateutil==2.9.0.post0
radius==0.0.4
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
